src/requirements.txt
boto3
rsa
pyjwt
attrs
requests
structlog
